Request a timer instead of claim and only init 4 GPIOS if in a 4 bit mode
Changed the code for FlexIO that instead of assuming Timer 2, it requests any free timer... So far it appears to work.

Also when in 4 bit mode only init the 4 data pins.  Before it was going off some specific  init mode.  Now instead check if G4 is 0xff and only do 4