Enable FrameBuffer on F/V free up memory
The previous code never enabled Frame buffer, and the
current code did not have enough memory left to actually
create the frame buffer.

So for now I updated the code to not statically allocate that large array for
saving to SD Card when that option is turned off, so now the set Frame buffer works...

The code is setup currently with two #ifdef options
Either to draw on Camera frame complete or on
TFT frame complete.

There is some tearing or like issue when moving.  That's next...

Warning I rebased this branch to start off from current main and then
had to fix a few conflicts.  So you should check to make
sure I did not screw it up!